+++ cinemabe/Cinema/3-Data/Cinema.Data.Entities/Entities/Discount.cs
@@ -3,8 +3,8 @@ public class Discount : BaseEntity
 {
     public string Code { get; set; } = string.Empty;
     public string? Description { get; set; }
-    public decimal Percent { get; set; }
-    public decimal? MaxDiscountAmount { get; set; }
+    public double Percent { get; set; }
+    public double? MaxDiscountAmount { get; set; }
     public Guid DiscountTypeId { get; set; }
     public DateTime StartDate { get; set; }
     public DateTime EndDate { get; set; }
