+++ cinemabe/Cinema/3-Data/Cinema.Data.Entities/Entities/Holiday.cs
@@ -3,5 +3,5 @@ public class Holiday : BaseEntity
 {
     public string Name { get; set; } = string.Empty;
     public DateOnly Date { get; set; }
-    public decimal PriceMultiplier { get; set; } = 1.5m;
+    public double PriceMultiplier { get; set; } = 1.5;
 }
