+++ cinemabe/Cinema/3-Data/Cinema.Data.Entities/Entities/InvoiceFoodAndDrink.cs
@@ -4,8 +4,8 @@ public class InvoiceFoodAndDrink
     public Guid InvoiceId { get; set; }
     public Guid FoodAndDrinkId { get; set; }
     public int Quantity { get; set; }
-    public decimal UnitPrice { get; set; }
-    public decimal TotalPrice { get; set; }
+    public double UnitPrice { get; set; }
+    public double TotalPrice { get; set; }
     public Invoice Invoice { get; set; } = null!;
     public FoodAndDrink FoodAndDrink { get; set; } = null!;
 }
